Additional cleanups - formatting, syntax, logging, moving code

+++ nucleus/cluster/admin/src/main/java/com/sun/enterprise/v3/admin/cluster/NodeUtils.java
@@ -28,8 +28,7 @@ import com.sun.enterprise.util.cluster.RemoteType;
 import com.sun.enterprise.util.net.NetUtils;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
+import java.lang.System.Logger;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.nio.file.Path;
@@ -37,7 +36,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.logging.Logger;
 
 import org.glassfish.api.ActionReport;
 import org.glassfish.api.admin.AdminCommandContext;
@@ -45,11 +43,14 @@ import org.glassfish.api.admin.CommandValidationException;
 import org.glassfish.api.admin.ParameterMap;
 import org.glassfish.api.admin.SSHCommandExecutionException;
 import org.glassfish.cluster.ssh.connect.NodeRunner;
+import org.glassfish.cluster.ssh.launcher.SSHException;
 import org.glassfish.cluster.ssh.launcher.SSHLauncher;
+import org.glassfish.common.util.admin.AuthTokenManager;
 import org.glassfish.hk2.api.ServiceLocator;
 import org.glassfish.internal.api.RelativePathResolver;
 
-import static java.util.logging.Level.WARNING;
+import static java.lang.System.Logger.Level.INFO;
+import static java.lang.System.Logger.Level.WARNING;
 
 /**
  * Utility methods for operating on Nodes
@@ -58,6 +59,8 @@ import static java.util.logging.Level.WARNING;
  * @author Byron Nevins
  */
 public class NodeUtils {
+    private static final Logger LOG = System.getLogger(NodeUtils.class.getName());
+
     public static final String NODE_DEFAULT_SSH_PORT = "22";
     public static final String NODE_DEFAULT_REMOTE_USER = "${user.name}";
     static final String NODE_DEFAULT_INSTALLDIR = "${com.sun.aas.productRoot}";
@@ -77,12 +80,10 @@ public class NodeUtils {
     static final Path LANDMARK_FILE = Path.of("glassfish", "modules", "admin-cli.jar");
     private static final String NL = System.lineSeparator();
     private TokenResolver resolver = null;
-    private Logger logger = null;
-    private ServiceLocator habitat = null;
+    private ServiceLocator locator = null;
 
-    NodeUtils(ServiceLocator habitat, Logger logger) {
-        this.logger = logger;
-        this.habitat = habitat;
+    NodeUtils(ServiceLocator locator) {
+        this.locator = locator;
 
         // Create a resolver that can replace system properties in strings
         Map<String, String> systemPropsMap =
@@ -111,7 +112,7 @@ public class NodeUtils {
         command.add("version");
         command.add("--local");
         command.add("--terse");
-        NodeRunner nr = new NodeRunner(habitat);
+        NodeRunner nr = new NodeRunner(locator.getService(AuthTokenManager.class));
 
         StringBuilder output = new StringBuilder();
         try {
@@ -284,16 +285,10 @@ public class NodeUtils {
         SSHLauncher sshL = new SSHLauncher(node);
         try {
             sshL.pingConnection();
-        } catch (Exception e) {
-            String m1 = e.getMessage();
-            String m2 = "";
-            Throwable e2 = e.getCause();
-            if (e2 != null) {
-                m2 = e2.getMessage();
-            }
-            String msg = Strings.get("ssh.bad.connect", node.getNodeHost(), "SSH");
-            logger.log(WARNING, StringUtils.cat(": ", msg, m1, m2), e);
-            throw new CommandValidationException(StringUtils.cat(NL, msg, m1, m2), e);
+        } catch (SSHException e) {
+            String msg = Strings.get("ssh.bad.connect", node.getNodeHost(), "SSH", e.getMessage());
+            LOG.log(WARNING, msg, e);
+            throw new CommandValidationException(msg, e);
         }
     }
 
@@ -337,25 +332,14 @@ public class NodeUtils {
             sshpassword, keyFile == null ? null : new File(keyFile), sshkeypassphrase);
         try {
             Path pathToCheck = resolvedInstallDir.resolve(LANDMARK_FILE);
-            if (!sshLauncher.exists(pathToCheck)) {
-                throw new FileNotFoundException(
+            if (!installFlag && !sshLauncher.exists(pathToCheck)) {
+                throw new CommandValidationException(
                     "Invalid install directory: could not find " + pathToCheck + " on " + host);
             }
-        } catch (IOException e) {
-            String m1 = e.getMessage();
-            String m2 = "";
-            Throwable e2 = e.getCause();
-            if (e2 != null) {
-                m2 = e2.getMessage();
-            }
-            if (e instanceof FileNotFoundException) {
-                if (!installFlag) {
-                    throw new CommandValidationException(StringUtils.cat(NL, m1, m2), e);
-                }
-            } else {
-                String msg = Strings.get("ssh.bad.connect", nodehost, "SSH");
-                throw new CommandValidationException(StringUtils.cat(NL, msg, m1, m2), e);
-            }
+        } catch (SSHException e) {
+            String msg = Strings.get("ssh.bad.connect", nodehost, "SSH", e.getMessage());
+            LOG.log(WARNING, msg, e);
+            throw new CommandValidationException(msg, e);
         }
     }
 
@@ -418,62 +402,48 @@ public class NodeUtils {
         }
 
         if (StringUtils.ok(humanCommand)) {
-            msg3 = Strings.get("node.remote.tocomplete",
-                    nodeHost, installDir, humanCommand);
+            msg3 = Strings.get("node.remote.tocomplete", nodeHost, installDir, humanCommand);
         }
 
-        NodeRunner nr = new NodeRunner(habitat);
+        NodeRunner nr = new NodeRunner(locator.getService(AuthTokenManager.class));
         try {
             int status = nr.runAdminCommandOnNode(node, output, command, context);
-            if (status != 0) {
-                // Command ran, but didn't succeed. Log full information
-                msg2 = Strings.get("node.command.failed", nodeName,
-                        nodeHost, output.toString().trim(), nr.getLastCommandRun());
-                logger.warning(StringUtils.cat(": ", msg1, msg2, msg3));
-                // Don't expose command name to user in case it is a hidden command
-                msg2 = Strings.get("node.command.failed.short", nodeName,
-                        nodeHost, output.toString().trim());
-            }
-            else {
+            if (status == 0) {
                 failure = false;
-                logger.info(output.toString().trim());
+                LOG.log(INFO, output.toString().trim());
+            } else {
+                // Command ran, but didn't succeed. Log full information
+                msg2 = Strings.get("node.command.failed", nodeName, nodeHost, output.toString().trim(),
+                    nr.getLastCommandRun());
+                LOG.log(WARNING, StringUtils.cat(": ", msg1, msg2, msg3));
+                // Don't expose command name to user in case it is a hidden command
+                msg2 = Strings.get("node.command.failed.short", nodeName, nodeHost, output.toString().trim());
             }
-        }
-        catch (SSHCommandExecutionException ec) {
-            msg2 = Strings.get("node.ssh.bad.connect", nodeName, nodeHost, ec.getMessage());
+        } catch (SSHCommandExecutionException e) {
+            msg2 = Strings.get("node.ssh.bad.connect", nodeName, nodeHost, e.getMessage());
             // Log some extra info
-            String msg = Strings.get("node.command.failed.ssh.details",
-                    nodeName, nodeHost, ec.getCommandRun(), ec.getMessage(),
-                    ec.getSSHSettings());
-            logger.log(WARNING, StringUtils.cat(": ", msg1, msg, msg3), ec);
-        }
-        catch (ProcessManagerException ex) {
-            msg2 = Strings.get("node.command.failed.local.details",
-                    ex.getMessage(), nr.getLastCommandRun());
-            logger.warning(StringUtils.cat(": ", msg1, msg2, msg3));
+            String msg = Strings.get("node.command.failed.ssh.details", nodeName, nodeHost, e.getCommandRun(),
+                e.getMessage(), e.getSSHSettings());
+            LOG.log(WARNING, StringUtils.cat(": ", msg1, msg, msg3), e);
+        } catch (ProcessManagerException e) {
+            msg2 = Strings.get("node.command.failed.local.details", e.getMessage(), nr.getLastCommandRun());
+            LOG.log(WARNING, StringUtils.cat(": ", msg1, msg2, msg3), e);
             // User message doesn't have command that was run
-            msg2 = Strings.get("node.command.failed.local.exception",
-                    ex.getMessage());
-        }
-        catch (UnsupportedOperationException e) {
+            msg2 = Strings.get("node.command.failed.local.exception", e.getMessage());
+        } catch (UnsupportedOperationException e) {
             msg2 = Strings.get("node.not.ssh", nodeName, nodeHost);
-            logger.warning(StringUtils.cat(": ", msg1, msg2, msg3));
-        }
-        catch (IllegalArgumentException e) {
+            LOG.log(WARNING, StringUtils.cat(": ", msg1, msg2, msg3), e);
+        } catch (IllegalArgumentException e) {
             msg2 = e.getMessage();
-            logger.warning(StringUtils.cat(": ", msg1, msg2, msg3));
+            LOG.log(WARNING, StringUtils.cat(": ", msg1, msg2, msg3), e);
         }
 
         if (failure) {
             report.setMessage(StringUtils.cat(NL + NL, msg1, msg2, msg3));
             report.setActionExitCode(ActionReport.ExitCode.FAILURE);
-        }
-        else {
+        } else {
             report.setActionExitCode(ActionReport.ExitCode.SUCCESS);
         }
-
-
-        return;
     }
 
     private RemoteType parseType(ParameterMap map) throws CommandValidationException {
